Move NaN-propagation tests back to burn-flex (backend-specific semantics)
The NaN-propagation tests for max_dim/min_dim/max_dim_with_indices/
argmax/cummin/cummax fail on ndarray because IEEE 754 leaves
min(NaN, x) implementation-defined: ndarray returns the non-NaN arg,
flex propagates NaN. These assertions only hold on flex.

Revert the move for:
- test_max_dim_nan_propagation
- test_min_dim_nan_propagation
- test_max_dim_with_indices_nan_propagation
- test_argmax_nan_propagation
- test_cummin_nan_propagation
- test_cummax_nan_propagation
- test_cummin_nan_at_start

Reinstated in burn-flex/src/ops/reduce.rs and cumulative.rs with the
flex-internal-call pattern (Flex::float_{max,min,argmax}_dim /
cummin_f32 / cummax_f32) and a leading comment explaining why the tests
stay flex-specific. Deleted the backend-tests copies.

Verified: cargo test -p burn-backend-tests --features ndarray,std --test tensor
passes 1490 tests (was 7 failing); flex and flex-backend suites still green.
